Update application to latest example-project & rename it to 'main'
