--- {{cookiecutter.project_name_slug}}/applications/app/app_version.h.in
@@ -1,29 +1,0 @@
-/**
- * @file app_version.h
- *
- * Application version information.
- *
- * Copyright (c) 2021 Nordic Semiconductor ASA
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#ifndef APP_VERSION_H_
-#define APP_VERSION_H_
-
-/** Application major version. */
-#define APP_VERSION_MAJOR ${PROJECT_VERSION_MAJOR}
-/** Application minor version. */
-#define APP_VERSION_MINOR ${PROJECT_VERSION_MINOR}
-/** Application patch version. */
-#define APP_VERSION_PATCH ${PROJECT_VERSION_PATCH}
-
-/** Application version. */
-#define APP_VERSION \
-	((APP_VERSION_MAJOR << 16) + \
-	 (APP_VERSION_MINOR << 8) + \
-	  APP_VERSION_PATCH)
-
-/** Application version (string). */
-#define APP_VERSION_STR "${PROJECT_VERSION}"
-
-#endif /* APP_VERSION_H_ */
