--- {{cookiecutter.project_name_slug}}/applications/app/rtt.conf
@@ -1,9 +1,0 @@
-# Copyright (c) 2021 Nordic Semiconductor ASA
-# SPDX-License-Identifier: Apache-2.0
-#
-# This is a Kconfig fragment which can be used to enable Segger RTT options.
-# It should be used in conjunction with debug.conf.
-
-# segger RTT console
-CONFIG_USE_SEGGER_RTT=y
-CONFIG_RTT_CONSOLE=y
