+++ {{cookiecutter.project_name_slug}}/applications/main/CMakeLists.txt
@@ -1,15 +1,13 @@
 #-------------------------------------------------------------------------------
 # Zephyr Example Application
 #
+# Copyright (c) {% now 'utc', '%Y' %} {{cookiecutter.project_org}}
 # Copyright (c) 2021 Nordic Semiconductor ASA
 # SPDX-License-Identifier: Apache-2.0
 
 cmake_minimum_required(VERSION 3.13.1)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 
-project(app LANGUAGES C VERSION 1.0.0)
-
-configure_file(app_version.h.in ${CMAKE_BINARY_DIR}/app/include/app_version.h)
-target_include_directories(app PRIVATE ${CMAKE_BINARY_DIR}/app/include src)
+project(app LANGUAGES C)
 
 target_sources(app PRIVATE src/main.c)
