+++ {{cookiecutter.project_name_slug}}/applications/main/Kconfig
@@ -1,3 +1,4 @@
+# Copyright (c) {% now 'utc', '%Y' %} {{cookiecutter.project_org}}
 # Copyright (c) 2021 Nordic Semiconductor ASA
 # SPDX-License-Identifier: Apache-2.0
 #
