+++ {{cookiecutter.project_name_slug}}/applications/main/VERSION
@@ -1,0 +1,5 @@
+VERSION_MAJOR = 1
+VERSION_MINOR = 0
+PATCHLEVEL = 0
+VERSION_TWEAK = 0
+EXTRAVERSION =
