+++ {{cookiecutter.project_name_slug}}/applications/main/debug.conf
@@ -1,3 +1,4 @@
+# Copyright (c) {% now 'utc', '%Y' %} {{cookiecutter.project_org}}
 # Copyright (c) 2021 Nordic Semiconductor ASA
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -6,13 +7,6 @@
 
 # compiler
 CONFIG_DEBUG_OPTIMIZATIONS=y
-
-# console
-CONFIG_CONSOLE=y
-
-# UART console
-CONFIG_SERIAL=y
-CONFIG_UART_CONSOLE=y
 
 # logging
 CONFIG_LOG=y
