+++ {{cookiecutter.project_name_slug}}/applications/main/prj.conf
@@ -1,6 +1,7 @@
+# Copyright (c) {% now 'utc', '%Y' %} {{cookiecutter.project_org}}
 # Copyright (c) 2021 Nordic Semiconductor ASA
 # SPDX-License-Identifier: Apache-2.0
 #
 # This file contains selected Kconfig options for the application.
 
-CONFIG_SENSOR=y
+CONFIG_BLINK=y
